Move FastaSequenceReader from eva-accession-pipeline to eva-accession-core
