+++ eva-accession-core/src/test/resources/input-files/fasta/Gallus_gallus-5.0.test.fa.dict
