+++ eva-accession-core/src/test/resources/input-files/fasta/compressed.fa.gz
