+++ eva-accession-core/src/test/resources/input-files/fasta/fastaWithNoDictionary.fa
