+++ eva-accession-core/src/test/resources/input-files/fasta/fastaWithNoIndex.fa
