+++ eva-accession-pipeline/src/main/java/uk/ac/ebi/eva/accession/pipeline/configuration/AccessionWriterConfiguration.java
@@ -23,7 +23,7 @@ import uk.ac.ebi.eva.accession.core.SubmittedVariantAccessioningService;
 import uk.ac.ebi.eva.accession.core.configuration.SubmittedVariantAccessioningConfiguration;
 import uk.ac.ebi.eva.accession.pipeline.io.AccessionReportWriter;
 import uk.ac.ebi.eva.accession.pipeline.io.AccessionWriter;
-import uk.ac.ebi.eva.accession.pipeline.io.FastaSequenceReader;
+import uk.ac.ebi.eva.accession.core.fasta.FastaSequenceReader;
 import uk.ac.ebi.eva.accession.pipeline.parameters.InputParameters;
 
 import java.io.File;
